+++ projects/angular-sdk/src/components/cfar-offer-banner-large/cfar-offer-banner-large.component.ts
@@ -1,14 +1,11 @@
-import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { take } from 'rxjs/operators';
-import { CfarContract, CfarItinerary, CfarOffer, CreateCfarOfferRequest, RequestType } from '../../apis/hopper-cloud-airline/v1';
+import { CfarContract, CfarItinerary, CfarOffer, CreateCfarContractRequest, CreateCfarOfferRequest, RequestType } from '../../apis/hopper-cloud-airline/v1';
 import { AbstractComponent } from '../abstract.component';
 import { TranslateService } from '@ngx-translate/core';
 import { DateAdapter } from "@angular/material/core";
 import { ApiTranslatorUtils } from '../../utils/api-translator.utils';
 import { HopperProxyService } from '../../services/hopper-proxy.service';
-import { CfarOfferDialogComponent, DialogUtils } from '../../public-api';
-import { MatDialog } from '@angular/material/dialog';
-import { MatRadioButton, MatRadioGroup } from '@angular/material/radio';
 
 @Component({
   selector: 'hopper-cfar-offer-banner-large',
@@ -17,14 +14,14 @@ import { MatRadioButton, MatRadioGroup } from '@angular/material/radio';
 })
 export class CfarOfferBannerLargeComponent extends AbstractComponent implements OnInit {
 
-  public cheapestOffer!: CfarOffer;
+  public cfarOffers!: CfarOffer[];
+  public selectedCfarOffer!: CfarOffer;
   public isLoading!: boolean;
-
-  private _cfarOffers!: CfarOffer[];
+  public selectedChoice!: number;
 
   @Input() hCSessionId!: string;
   @Input() itineraries!: CfarItinerary[];
-  @Input() currentTheme!: string;
+  @Input() pnrReference!: string;
 
   @Output() offerAccepted = new EventEmitter();
 
@@ -32,9 +29,11 @@ export class CfarOfferBannerLargeComponent extends AbstractComponent implements 
     private _adapter: DateAdapter<any>,
     private _translateService: TranslateService,
     private _hopperProxyService: HopperProxyService,
-    private _dialog: MatDialog
   ) {
     super(_adapter, _translateService);
+
+    // Decline CFAR option, by default
+    this.selectedChoice = -1;
   }
 
   // -----------------------------------------------
@@ -43,8 +42,8 @@ export class CfarOfferBannerLargeComponent extends AbstractComponent implements 
 
   ngOnInit(): void {
     if (this.isFakeBackend) {
-      this._cfarOffers = this._buildFakePostCfarOffersResponse();
-      this.cheapestOffer = this._getCheapestOffer(this._cfarOffers);
+      this.cfarOffers = this._buildFakePostCfarOffersResponse();
+      this.selectedCfarOffer = this._getCheapestOffer(this.cfarOffers);
     } else {
       this.isLoading = true;
 
@@ -61,8 +60,8 @@ export class CfarOfferBannerLargeComponent extends AbstractComponent implements 
               });
             }
             
-            this._cfarOffers = results;
-            this.cheapestOffer = this._getCheapestOffer(this._cfarOffers);
+            this.cfarOffers = results;
+            this.selectedCfarOffer = this._getCheapestOffer(this.cfarOffers);
             this.isLoading = false;
           },
           (error) => this.isLoading = false
@@ -74,29 +73,38 @@ export class CfarOfferBannerLargeComponent extends AbstractComponent implements 
   // Publics Methods
   // -----------------------------------------------
 
-  public onSubmit(cfarOffer: CfarOffer): void {
-    const dialogData = { 
-      currentLang: this.currentLang,
-      basePath: this.basePath,
-      hCSessionId: this.hCSessionId,
-      cfarOffers: this._cfarOffers,
-      extAttributes: this.extAttributes
-    };
-    const dialogConfig = DialogUtils.getDialogConfig(dialogData, this.currentTheme);
-    const dialogRef = this._dialog.open(CfarOfferDialogComponent, dialogConfig);
+  public onSubmit(): void {
+    if (this.selectedChoice != -1) {
+      this.isLoading = true;
 
-    dialogRef.afterClosed()
-      .pipe(take(1))
-      .subscribe(
-        (result: CfarContract) => {
-          if (result) {
-            this.offerAccepted.emit(result);
-          } else {
-            console.log("Close dialog");
-          }
-        },
-        (error) => console.log(error)
-      );
+      // Create CFAR Contract
+      this._hopperProxyService
+        .postCfarContracts(this.basePath, this.hCSessionId, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarContractRequest()))
+        .pipe(take(1))
+        .subscribe(
+          (cfarContract: CfarContract) => {
+            this.offerAccepted.emit(cfarContract);
+            this.isLoading = false;
+          },
+          (error) => this.isLoading = false
+        );
+    }
+  }
+
+  public computePercentage(offer: CfarOffer): number {
+    if (offer) {
+      const coverage = Number.parseFloat(offer.coverage);
+      const totalPrice = Number.parseFloat(offer.itinerary.totalPrice);
+  
+      return coverage / (totalPrice || 1.0);
+    }
+
+    return 0;
+  }
+
+  public onChangeChoice(): void {
+    // Update descriptions
+    this.selectedCfarOffer = this.selectedChoice > -1 ? this.cfarOffers[this.selectedChoice] : this._getCheapestOffer(this.cfarOffers);
   }
 
   // -----------------------------------------------
@@ -110,6 +118,15 @@ export class CfarOfferBannerLargeComponent extends AbstractComponent implements 
       extAttributes: this.extAttributes
     };
   }
+
+  private _buildCreateCfarContractRequest(): CreateCfarContractRequest {
+    return {
+      offerIds: [this.selectedCfarOffer.id],
+      itinerary: this.selectedCfarOffer.itinerary,
+      extAttributes: this.extAttributes,
+      pnrReference: this.pnrReference
+    };
+  } 
 
   private _buildFakePostCfarOffersResponse(): CfarOffer[] {   
     return [